Added code to recieve SAME messages and display them.
Changed the eaxmple code mainly modifying the loop to updated SAME
message and a valid channel light.  It also automatically recieves the
SAME message and sends it up to the serial port.

Added clear_SAME_INTACK to acknowledge the SAME message.

Added command_SAME_Message to retrieve the SAME message and return it.